requirements.txt
alembic==1.0.11
certifi==2019.6.16
chardet==3.0.4
Click==7.0
dominate==2.4.0
Flask==1.1.1
Flask-Bootstrap==3.3.7.1
Flask-Login==0.4.1
Flask-Migrate==2.5.2
Flask-Moment==0.9.0
Flask-SQLAlchemy==2.4.0
Flask-Uploads==0.2.1
Flask-WTF==0.14.2
gevent==1.4.0
greenlet==0.4.15
idna==2.8
itsdangerous==1.1.0
Jinja2==2.10.1
Mako==1.1.0
MarkupSafe==1.1.1
oauthlib==3.1.0
PIL==1.1.7
Pillow==6.1.0
pyasn1==0.4.7
pyasn1-modules==0.2.6
python-dateutil==2.8.0
python-dotenv==0.10.3
python-editor==1.0.4
python-ldap==3.2.0
python-thumbnails==0.5.1
requests==2.6.0
requests-oauthlib==1.2.0
six==1.9.0
SQLAlchemy==1.3.7
urllib3==1.25.3
visitor==0.1.3
Wand==0.5.6
Werkzeug==0.15.5
WTForms==2.2.1
